Custom hook created to manage autocomplete inputs
