+++ src/pages/genre/more-genres/mobile-view.tsx
@@ -1,52 +1,19 @@
-import { useEffect, useRef, useState } from "react";
 import { GenreBookListItemsMobile } from "../../../components/genres/list-items-mobile"
 import { MoreGenresSelect } from "../../../components/genres/more-genres-select";
 import { GenresSearchBarMobile } from "../../../components/genres/search-bar-mobile"
 import { Loading } from "../../../components/loading";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
+import { useAppSelector } from "../../../redux/hooks";
 import type { RootState } from "../../../redux/store";
 import { AutocompleteInput } from "../../../components/autocomplete";
 import { GenreAutocompleteItem } from "../../../components/autocomplete/genre-results-item";
 import genresActions from "../../../redux/actions/genres";
-import debounce from "lodash.debounce";
+import { useAutocomplete } from "../../../hooks/autocomplete";
 
 
 export const MoreGenresMobile = () => {
-    const dispatch = useAppDispatch();
     const { genresList, discoverList, searchResults, status: genresStatus } = useAppSelector((state: RootState) => state.genres);
-    const [searchValue, setSearchValue] = useState("");
-    const [isAutocompleteOpen, setIsAutocompleteOpen] = useState(false);
-    const autocompleteRef = useRef<HTMLDivElement>(null);
+    const { autocompleteRef, debouncedHandleOnChangeSearch, searchValue, isAutocompleteOpen } = useAutocomplete(genresActions.searchByName);
 
-    const handleOnChangeSearch = async (value: string) => {
-        setSearchValue(value);
-        await dispatch(genresActions.searchByName({ query: value }));
-
-        if (!isAutocompleteOpen && value) {
-            setIsAutocompleteOpen(true)
-        } else if (isAutocompleteOpen && !value) {
-            setIsAutocompleteOpen(false)
-        }
-    }
-
-    const debouncedHandleOnChangeSearch = debounce(handleOnChangeSearch, 100);
-
-    useEffect(() => {
-        const handleClickOutside = (event: MouseEvent) => {
-            if (
-                autocompleteRef.current &&
-                !autocompleteRef.current.contains(event.target as Node)
-            ) {
-                setIsAutocompleteOpen(false);
-            }
-        }
-
-        document.addEventListener("mousedown", handleClickOutside);
-
-        return () => {
-            document.removeEventListener("mousedown", handleClickOutside);
-        };
-    }, [])
 
     if (genresStatus === "loading") {
         return <Loading />
